Extract `StandardMetadataBlock` in dedicated file
