+++ lib/fastlane/plugin/wpmreleasetoolkit/helper/metadata_update_helper.rb
@@ -1,49 +1,8 @@
 require_relative 'metadata_block'
+require_relative 'standard_metadata_block'
 
 module Fastlane
   module Helper
-    class StandardMetadataBlock < MetadataBlock
-      attr_reader :content_file_path
-
-      def initialize(block_key, content_file_path)
-        super(block_key)
-        @content_file_path = content_file_path
-      end
-
-      def is_handler_for(key)
-        key == @block_key.to_s
-      end
-
-      def handle_line(fw, line)
-        # put the new content on block start
-        # and skip all the other content
-        generate_block(fw) if line.start_with?('msgctxt')
-      end
-
-      def generate_block(fw)
-        # init
-        fw.puts("msgctxt \"#{@block_key}\"")
-        line_count = File.foreach(@content_file_path).inject(0) { |c, _line| c + 1 }
-
-        if line_count <= 1
-          # Single line output
-          fw.puts("msgid \"#{File.read(@content_file_path).rstrip}\"")
-        else
-          # Multiple line output
-          fw.puts('msgid ""')
-
-          # insert content
-          File.open(@content_file_path, 'r').each do |line|
-            fw.puts("\"#{line.strip}\\n\"")
-          end
-        end
-
-        # close
-        fw.puts('msgstr ""')
-        fw.puts('')
-      end
-    end
-
     class WhatsNewMetadataBlock < StandardMetadataBlock
       attr_reader :new_key, :old_key, :rel_note_key, :release_version
